BigQuery: Convert `json` atoms to a JSON column type if configured in `mozPipelineMetadata`
This allows the `object` metric type to become a JSON column in
BigQuery.

This is the _only_ place where jsonschema-transpiler uses
schema metadata specific to Mozilla for generating BigQuery schemas.

It's added to the context after parsing the schema into an AST.
This _could_ be a CLI parameter instead, but due to how
jsonschema-transpiler is used it is easier for us to pass that
information as configuration within the schema itself.
It's still generic enough that we could extend it to other fields in the
future if needed.