Construct full namespace when infering name
I'm decently sure this was intended from the beginning, but was always
wrong.

Notably we have a function `fully_qualified_name` which is NOT used
here.
That one at least documents the expected fully qualified name. However
it has a `root` prefix, that doesn't show up anywhere.
And it also sets name to `__unknown__` if it is empty.
The last part is why we are NOT using this function here, but re-do the
work again: We don't want that `__unknown__` as part of the namespace
here
(and I wanted this patch as minimal as possible without rewriting and
testing `fully_qualified_name` throughout)